+++ crates/re_data_ui/src/annotation_context.rs
@@ -5,7 +5,7 @@ use re_types::components::AnnotationContext;
 use re_types::datatypes::{
     AnnotationInfo, ClassDescription, ClassDescriptionMapElem, KeypointId, KeypointPair,
 };
-use re_ui::UiExt as _;
+use re_ui::{DesignTokens, UiExt as _};
 use re_viewer_context::{auto_color, UiLayout, ViewerContext};
 
 use super::DataUi;
@@ -177,7 +177,7 @@ fn class_description_ui(
         ui_layout = UiLayout::SelectionPanelFull;
     }
 
-    let row_height = re_ui::DesignTokens::table_line_height();
+    let row_height = DesignTokens::table_line_height();
     if !class.keypoint_annotations.is_empty() {
         ui.maybe_collapsing_header(
             use_collapsible,
@@ -211,8 +211,8 @@ fn class_description_ui(
                         .column(Column::auto().clip(true).at_least(40.0))
                         .column(Column::auto().clip(true).at_least(40.0));
                     table
-                        .header(re_ui::DesignTokens::table_header_height(), |mut header| {
-                            re_ui::DesignTokens::setup_table_header(&mut header);
+                        .header(DesignTokens::table_header_height(), |mut header| {
+                            DesignTokens::setup_table_header(&mut header);
                             header.col(|ui| {
                                 ui.strong("From");
                             });
@@ -221,7 +221,7 @@ fn class_description_ui(
                             });
                         })
                         .body(|mut body| {
-                            re_ui::DesignTokens::setup_table_body(&mut body);
+                            DesignTokens::setup_table_body(&mut body);
 
                             // TODO(jleibs): Helper to do this with caching somewhere
                             let keypoint_map: ahash::HashMap<KeypointId, AnnotationInfo> = {
@@ -268,7 +268,7 @@ fn annotation_info_table_ui(
 ) {
     re_tracing::profile_function!();
 
-    let row_height = re_ui::DesignTokens::table_line_height();
+    let row_height = DesignTokens::table_line_height();
 
     ui.spacing_mut().item_spacing.x = 20.0; // column spacing.
 
@@ -282,8 +282,8 @@ fn annotation_info_table_ui(
         .column(Column::auto()); // color
 
     table
-        .header(re_ui::DesignTokens::table_header_height(), |mut header| {
-            re_ui::DesignTokens::setup_table_header(&mut header);
+        .header(DesignTokens::table_header_height(), |mut header| {
+            DesignTokens::setup_table_header(&mut header);
             header.col(|ui| {
                 ui.strong("Class Id");
             });
@@ -295,7 +295,7 @@ fn annotation_info_table_ui(
             });
         })
         .body(|mut body| {
-            re_ui::DesignTokens::setup_table_body(&mut body);
+            DesignTokens::setup_table_body(&mut body);
 
             body.rows(row_height, annotation_infos.len(), |mut row| {
                 let info = &annotation_infos[row.index()];
@@ -332,7 +332,7 @@ fn color_ui(ui: &mut egui::Ui, info: &AnnotationInfo, size: Vec2) {
 }
 
 fn small_color_ui(ui: &mut egui::Ui, info: &AnnotationInfo) {
-    let size = egui::Vec2::splat(re_ui::DesignTokens::table_line_height());
+    let size = egui::Vec2::splat(DesignTokens::table_line_height());
 
     let color = info
         .color
